// src/types/generated/start-stop.ts
export interface operations {
    getStart: {
        parameters: {
            query: {
                issueUrl: string;
                userId: string;
                environment?: "development" | "production";
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Successful response */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        ok: boolean;
                        computed: {
                            deadline: unknown;
                            isTaskStale: boolean;
                            wallet: string;
                            toAssign: string[];
                            senderRole: string;
                            consideredCount: number;
                            assignedIssues: {
                                title: string;
                                html_url: string;
                            }[];
                        };
                        warnings: {
                            logMessage: {
                                level: string;
                                raw: string;
                                diff: string;
                                type: string;
                            };
                            metadata: Record<string, never>;
                        }[];
                        reasons: string[] | null;
                    };
                };
            };
        };
    };
    postStart: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": {
                    /** Format: uri */
                    issueUrl: string;
                    userId: string;
                    environment?: "development" | "production";
                };
            };
        };
        responses: {
            /** @description Successful response */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        ok: boolean;
                        computed: {
                            deadline: unknown;
                            isTaskStale: boolean;
                            wallet: string;
                            toAssign: string[];
                            senderRole: string;
                            consideredCount: number;
                            assignedIssues: {
                                title: string;
                                html_url: string;
                            }[];
                        };
                        warnings: {
                            logMessage: {
                                level: string;
                                raw: string;
                                diff: string;
                                type: string;
                            };
                            metadata: Record<string, never>;
                        }[];
                        reasons: string[] | null;
                    };
                };
            };
        };
    };
}